+++ tests/run/arithmetic_analyse_types.pyx
@@ -10,11 +10,11 @@ cdef extern from "arithmetic_analyse_types_helper.h":
         char *msg
     cdef struct longlong_return:
         char *msg
-    cdef short_return f(short)
-    cdef int_return f(int)
-    cdef longlong_return f(long long)
+    cdef short_return f(i16)
+    cdef int_return f(i32)
+    cdef longlong_return f(i128)
 
-def short_binop(short val):
+def short_binop(i16 val):
     """
     Arithmetic in C is always done with at least int precision.
     
@@ -27,7 +27,7 @@ def short_binop(short val):
     cdef int_return x = f(val + val)
     return x.msg.decode('ASCII')
 
-def short_unnop(short val):
+def short_unnop(i16 val):
     """
     Arithmetic in C is always done with at least int precision.
     
@@ -37,7 +37,7 @@ def short_unnop(short val):
     cdef int_return x = f(-val)
     return x.msg.decode('ASCII')
 
-def longlong_binop(long long val):
+def longlong_binop(i128 val):
     """
     >>> print(longlong_binop(3))
     long long called
@@ -45,14 +45,13 @@ def longlong_binop(long long val):
     cdef longlong_return x = f(val * val)
     return x.msg.decode('ASCII')
 
-def longlong_unnop(long long val):
+def longlong_unnop(i128 val):
     """
     >>> print(longlong_unnop(3))
     long long called
     """
     cdef longlong_return x = f(~val)
     return x.msg.decode('ASCII')
-
 
 def test_bint(bint a):
     """
